+++ py2app/bootstrap/boot_app.py
@@ -1,13 +1,14 @@
-def _run(*scripts):
+def _run():
     global __file__
     import os, sys, site
     sys.frozen = 'macosx_app'
     base = os.environ['RESOURCEPATH']
 
-    for script in scripts:
-        path = os.path.join(base, script)
-        sys.argv[0] = __file__ = path
-        with open(path) as fp:
-            source = fp.read() + "\n"
-        exec(compile(source, path, 'exec'), globals(), globals())
+    argv0 = os.path.basename(os.environ['ARGVZERO'])
+    script = SCRIPT_MAP.get(argv0, DEFAULT_SCRIPT)
 
+    path = os.path.join(base, script)
+    sys.argv[0] = __file__ = path
+    with open(path) as fp:
+        source = fp.read() + "\n"
+    exec(compile(source, path, 'exec'), globals(), globals())