+++ py2app_tests/app_with_scripts/helper2.py
@@ -1,0 +1,2 @@
+import code
+print("Helper 2")