+++ py2app_tests/plugin_with_scripts/helper1.py
@@ -1,0 +1,1 @@
+print("Helper 1")