+++ py2app_tests/plugin_with_scripts/main.py
@@ -1,0 +1,7 @@
+import sys
+import Foundation
+
+class BasicPlugin (Foundation.NSObject):
+    def performCommand_(self, cmd):
+        print ("+ %s"%(cmd,))
+        sys.stdout.flush()